Map is fixed

+++ Voice Note/Views/DetailView.swift
@@ -1,23 +1,29 @@
 import SwiftUI
 import CoreLocation
+import MapKit
 
 struct DetailView: View {
     @EnvironmentObject var voiceNoteViewModel: VoiceNoteViewModel
+    @EnvironmentObject var mapViewModel: MapViewModel
     @Environment(\.presentationMode) var presentationMode
+    @Environment(\.coreData) private var coreDataService: CoreDataService
     @State private var showShareSheet = false
     let textContainer = #colorLiteral(red: 0.4, green: 0.2039215686, blue: 0.4980392157, alpha: 0.2)
-    private let voiceNote: VoiceNote
+    private let voiceNote: VoiceNote?
     
     @State private var isEditing = false
     @State private var editText: String
     
-    init(voiceNote: VoiceNote) {
-        self.voiceNote = voiceNote
-        _editText = State(initialValue: voiceNote.text ?? "")
+    init(voiceNoteUUID uuid: UUID) {
+            voiceNote = CoreDataService.localStorage.getVoiceNote(byUUID: uuid)
+        _editText = State(initialValue: voiceNote?.text ?? "Title Missing")
     }
     
     var body: some View {
         VStack {
+            Text("\(voiceNote?.title ?? "Missing Title")")
+                    .font(.largeTitle)
+                    .padding(.top, 8)
             if isEditing {
                 TextEditor(text: $editText)
                     .padding()
@@ -26,7 +32,7 @@ struct DetailView: View {
                     .cornerRadius(20)
                     .padding()
             } else {
-                Text(voiceNote.text ?? "")
+                Text(voiceNote?.text ?? "")
                     .padding()
                     .frame(maxWidth: .infinity)
                     .background(Color(textContainer))
@@ -34,34 +40,39 @@ struct DetailView: View {
                     .padding()
             }
             
-            RecordingCardView(voiceNoteUrl: voiceNote.fileUrl).padding(15)
-            Text("Duration: \(voiceNote.duration)s")
+            RecordingCardView(voiceNoteUrl: voiceNote?.fileUrl).padding(8)
+            Text("Duration: \(TimeDuration(size: voiceNote?.duration ?? 0).getTimeAsHourMinuteSecond())")
             
            
             
             HStack {
                 // Direction button
-                DetailBtn(clickHander: {
+                DetailBtn(clickHandler: {
                     print("Direction pressed")
+                    guard let latitude = voiceNote?.location?.latitude,
+                          let longitude = voiceNote?.location?.longitude,
+                          let title = voiceNote?.title else { return }
+                    let destination = CLLocationCoordinate2D(latitude: latitude, longitude: longitude)
+                    mapViewModel.openDirectionInAppleMapFromCurrentLocation(to: destination, directionMode: MKLaunchOptionsDirectionsModeDriving, destinationName: title)
                 }, icon: "arrow.triangle.turn.up.right.diamond.fill")
                 
                 //  Edit button
-                DetailBtn(clickHander: {
+                DetailBtn(clickHandler: {
                     toggleEditing()
                 }, icon: isEditing ? "checkmark.circle.fill" : "pencil")
 
                 
                 // Delete button
-                DetailBtn(clickHander: {
+                DetailBtn(clickHandler: {
                     deleteVoiceNote()
                 }, icon: "trash")
                 
                 // Share button
-                DetailBtn(clickHander: {
+                DetailBtn(clickHandler: {
                     showShareSheet = true
                 }, icon: "square.and.arrow.up")
                 .sheet(isPresented: $showShareSheet, content: {
-                    ShareSheet(activityItems: [voiceNote.text ?? ""])
+                    ShareSheet(activityItems: [voiceNote?.text ?? ""])
                 })
             }
             .presentationDetents([.medium, .large])
@@ -84,22 +95,26 @@ struct DetailView: View {
         This method deletes the voice note from CoreData
      */
     private func deleteVoiceNote() {
-        let context = PersistenceController.shared.container.viewContext
-        context.delete(voiceNote)
-        do {
-            try context.save()
+        let context = coreDataService.getManageObjectContext()
+        if let voiceNote = voiceNote, let url = voiceNote.fileUrl {
+            //delete from DB
+            context.delete(voiceNote)
+            //delete local file
+            voiceNoteViewModel.deleteRecording(url: url)
+            // repopulate file URL array
+            voiceNoteViewModel.fetchVoiceNotes()
+            // repopulate map
+            mapViewModel.populateLocation()
             presentationMode.wrappedValue.dismiss()
-        } catch {
-            print("Error deleting voice note: \(error)")
         }
     }
     
     /**
-        This method update the editted text to CoreData
+        This method update the edited text to CoreData
      */
     private func saveEditedText() {
-        let context = PersistenceController.shared.container.viewContext
-        voiceNote.text = editText
+        let context = coreDataService.getManageObjectContext()
+        voiceNote?.text = editText
         do {
             try context.save()
         } catch {
@@ -109,12 +124,12 @@ struct DetailView: View {
     
     struct DetailBtn: View {
         let buttonColor = #colorLiteral(red: 0.1764705926, green: 0.01176470611, blue: 0.5607843399, alpha: 1)
-        let clickHander: () -> Void
+        let clickHandler: () -> Void
         let icon: String
         
         var body: some View {
             Button {
-                clickHander()
+                clickHandler()
             } label: {
                 Image(systemName: icon)
                     .font(.title2)
